Issue #17: Move should_respond logic to receiver.py
- Move SlackClient, is_bot_in_thread, should_respond to receiver
- Filter events at receiver level to reduce Worker Lambda invocations
- Remove redundant code from worker/handler.py
